Split slash command integration tests from workspace model
